requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.27.0
pydantic==2.11.7
python-dotenv==1.1.1

# Vector Store + Embeddings
faiss-cpu==1.9.0.post1
numpy==2.1.3
sentence-transformers

# LangChain - Orchestration Layer
langchain==0.3.25
langchain-google-genai==2.1.4
langchain-openai
langchain-groq
langchain-community==0.3.24
langchain-text-splitters==0.3.8

# Web Scraping - Kent Knowledge Base
beautifulsoup4==4.12.3
requests==2.32.3

# Optional PII NER (install model: python -m spacy download en_core_web_sm)
spacy>=3.7.2,<4.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
